cleanup of admin user and collector profile page
